refactor(frontend): convert `features/pages` to TypeScript (#82)
