+++ frontend/src/features/pages/index.ts
@@ -1,4 +1,4 @@
-export { default as HeroPage } from "./HeroPage"
-export { default as HomePage } from "./HomePage"
-export { default as LoggedOutPage } from "./LoggedOutPage"
-export { default as NotFoundPage } from "./NotFoundPage"
+export { default as HeroPage } from "./hero-page"
+export { default as HomePage } from "./home-page"
+export { default as LoggedOutPage } from "./logged-out-page"
+export { default as NotFoundPage } from "./not-found-page"
